New Save feature
Refactored the main player objects and methods in the same way that the
enemy constructor works. A new script was created player generator, and
the character script contains only the player object with attributes.
The save feature converts this player object into JSON data, and stored
in the localstorage feature of HTML5. The load feature retrieves the
JSON data from the local storage, and converts it into the player object
for use during the player generation method before every battle.